+++ Changelog.md
@@ -3,7 +3,7 @@
 ## [Unreleased]
 
 - [apple2] send the network unit number with the payloads for network calls to allow for multiple network sub-devices on a SP network device
-- [apple2] convert apple2 asm files to C for almost all network functions
+- [apple2] convert apple2 asm files to C for almost all network functions and bus calls
 
 ## [4.4.0] - 2024-07-28